rename the module name and add enum method
