+++ src/ai/backend/agent/observer/BUILD
